move to new featured workspace controller
