+++ api/src/main/java/org/pmiops/workbench/api/FeaturedWorkspacesController.java
@@ -11,25 +11,26 @@ import org.springframework.web.bind.annotation.RestController;
 @Deprecated(since = "July 2024", forRemoval = true)
 @RestController
 public class FeaturedWorkspacesController implements FeaturedWorkspacesConfigApiDelegate {
-  private final Provider<FeaturedWorkspacesConfig> configProvider;
+  private final Provider<FeaturedWorkspacesConfig> featuredWorkspacesConfigProvider;
 
   @Autowired
-  FeaturedWorkspacesController(Provider<FeaturedWorkspacesConfig> configProvider) {
-    this.configProvider = configProvider;
+  FeaturedWorkspacesController(
+      Provider<FeaturedWorkspacesConfig> featuredWorkspacesConfigProvider) {
+    this.featuredWorkspacesConfigProvider = featuredWorkspacesConfigProvider;
   }
 
   /**
    * @deprecated Get all Featured workspaces, using the Featured Workspaces Config JSON as the
    *     source. Use this in conjunction with {@link WorkspaceService#getPublishedWorkspaces()} to
    *     construct the list of (pre-July-2024) Featured workspaces.
-   *     <p>Use {@link WorkspaceService#getFeaturedWorkspaces()} to retrieve the Featured/Published
-   *     workspaces from July 2024 onward.
+   *     <p>Use {@link FeaturedWorkspaceController#getFeaturedWorkspaces()} to retrieve the
+   *     Featured/Published workspaces from July 2024 onward.
    * @return List of all Featured workspaces
    */
   @Deprecated(since = "July 2024", forRemoval = true)
   @Override
   public ResponseEntity<FeaturedWorkspacesConfigResponse> getFeaturedWorkspacesConfig() {
-    FeaturedWorkspacesConfig fwConfig = configProvider.get();
+    FeaturedWorkspacesConfig fwConfig = featuredWorkspacesConfigProvider.get();
     return ResponseEntity.ok(
         new FeaturedWorkspacesConfigResponse().featuredWorkspacesList(fwConfig.featuredWorkspaces));
   }
